refactor: cognitoの呼び出しをまとめた&リファクタした
