--- client/src/components/pages/Mypage/helper.ts
@@ -1,9 +1,0 @@
-import { CognitoUser } from "amazon-cognito-identity-js";
-
-export const signOutCognito = (cognitoUser: CognitoUser) => {
-  return new Promise<void>((res) => {
-    cognitoUser.signOut(() => {
-      res();
-    });
-  });
-};
